requirements.txt
boto3==1.24.89
botocore==1.27.89
greenlet==1.1.3
importlib-metadata==5.0.0
jmespath==1.0.1
kafka-python==2.0.2
mrjob==0.7.4
numpy==1.21.6
pandas==1.3.5
pkg_resources==0.0.0
psycopg2-binary==2.9.3
py4j==0.10.9.5
pyspark==3.3.0
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.41
sqlparse==0.4.3
typing_extensions==4.3.0
urllib3==1.26.12
zipp==3.8.1
